requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
numpy
matplotlib
yfinance
scipy
pandas
pyyaml
flask
curl_cffi
gunicorn
